+++ metainfo/evasion_attack_parameters.json
@@ -11,6 +11,12 @@
   "perturb_structure": ["perturb_structure", "bool", true, {}, "Indicates whether the structure can be changed"],
   "direct": ["direct", "bool", true, {}, "Indicates whether to directly modify edges/features of the node attacked or only those of influencers"],
   "n_influencers": ["n_influencers", "int", 0, {"min":  0, "step": 1}, "Number of influencing nodes. Will be ignored if direct is True"]
+ },
+  "PGD": {
+  "perturb_ratio": ["Perturbation ratio", "float", 0.5, {"min": 0, "max": 1, "step": 0.01}, "Ratio of edges subject to perturbation"],
+  "learning_rate": ["Learning rate", "float", 0.01, {}, "Learning rate for adjacency matrix optimization"],
+  "num_iterations": ["Number of iterations", "int", 100, {"min":  1, "step": 1}, "Number of iterations of gradient descent"],
+  "num_rand_trials": ["Number of random trials", "int", 100, {"min":  1, "step": 1}, "number of random trials in Random Sampling Algorithm"]
  }
 }
