+++ metainfo/poison_attack_parameters.json
@@ -15,12 +15,6 @@
   "train_iters": ["Train iters (surrogate)", "int", 200, {"min": 0, "step": 1}, "Trainig iterations for surrogate model"],
   "attack_structure": ["Attack structure", "bool", true, {}, "whether change graph structure with attack or not"],
   "attack_features": ["Attack features", "bool", false, {}, "whether change node features with attack or not"]
- },
- "PGD": {
-  "perturb_ratio": ["Perturbation ratio", "float", 0.5, {"min": 0, "max": 1, "step": 0.01}, "Ratio of edges subject to perturbation"],
-  "learning_rate": ["Learning rate", "float", 0.01, {}, "Learning rate for adjacency matrix optimization"],
-  "num_iterations": ["Number of iterations", "int", 100, {"min":  1, "step": 1}, "Number of iterations of gradient descent"],
-  "num_rand_trials": ["Number of random trials", "int", 100, {"min":  1, "step": 1}, "number of random trials in Random Sampling Algorithm"]
  }
 }
