+++ src/attacks/evasion_attacks.py
@@ -8,6 +8,13 @@ from attacks.attack_base import Attacker
 from src.attacks.nettack.nettack import Nettack
 from src.attacks.nettack.utils import preprocess_graph, largest_connected_components, data_to_csr_matrix, train_w1_w2
 
+# PGD imports
+from attacks.evasion_attacks_collection.pgd.utils import Projection, RandomSampling
+from attacks.evasion_attacks_collection.pgd.model import MatGNN
+import torch.nn.functional as F
+from torch_geometric.utils import to_dense_adj, dense_to_sparse
+from tqdm import tqdm
+
 
 class EvasionAttacker(Attacker):
     def __init__(self, **kwargs):
@@ -40,6 +47,104 @@ class FGSMAttacker(EvasionAttacker):
         perturbed_data_x = torch.clamp(perturbed_data_x, 0, 1)
         gen_dataset.data.x = perturbed_data_x.detach()
         return gen_dataset
+
+
+class PGDAttacker(EvasionAttacker):
+    name = "PGD"
+
+    def __init__(self,
+                 perturb_ratio=0.5,
+                 learning_rate=0.01,
+                 num_iterations=100,
+                 num_rand_trials=100):
+
+        super().__init__()
+        self.attack_diff = None
+        self.perturb_ratio = perturb_ratio
+        self.learning_rate = learning_rate
+        self.num_iterations = num_iterations
+        self.num_rand_trials = num_rand_trials
+
+    def attack(self, model_manager, gen_dataset, mask_tensor):
+        # Since the PGD attack is an attack on the graph structure, which requires optimization of the graph adjacency
+        # matrix, we need to use the GCNConv graph model, implemented through matrix operations with the ability to
+        # differentiate the adjacency matrix.
+        hidden = model_manager.gnn.structure.layers[0]['layer']['layer_kwargs']['out_channels']
+        model = MatGNN(num_features=gen_dataset.num_node_features, hidden=hidden, num_classes=gen_dataset.num_classes)
+
+        # Copy learned matrix
+        with torch.no_grad():
+            model.conv0.linear.weight.copy_(model_manager.gnn.GCNConv_0.lin.weight)  # W0
+            model.conv0.linear.bias.copy_(model_manager.gnn.GCNConv_0.bias)  # b0
+            model.conv1.linear.weight.copy_(model_manager.gnn.GCNConv_1.lin.weight)  # W1
+            model.conv1.linear.bias.copy_(model_manager.gnn.GCNConv_1.bias)  # b1
+
+        # Convert the list of edges into an adjacency matrix
+        data = gen_dataset.data
+        adj_matrix = to_dense_adj(data.edge_index).squeeze(0)
+
+        # eps - number of edges subject to perturbation
+        total_edges = data.edge_index.size(1)
+        eps = int(self.perturb_ratio * total_edges)
+
+        # Adjacency Matrix Optimization Process
+        # --------------- Start ---------------
+        A = adj_matrix
+        N = data.x.size(0)
+
+        # M - training mask matrix
+        M = torch.zeros((N, N), requires_grad=True)
+
+        # Projection operator
+        projection = Projection(eps=eps)
+
+        model.eval()
+        optimizer = torch.optim.Adam([M], lr=self.learning_rate, weight_decay=5e-4)
+
+        # Optimization cycle
+        progress_bar = tqdm(range(self.num_iterations), desc="Optimization cycle", leave=True, postfix={"Loss": 0.0})
+        for t in progress_bar:
+            # Perturbation of matrix A; A_pert is the perturbed matrix
+            A_pert = A - A * M
+            preds = model(data.x, A_pert)
+
+            # calculate the loss
+            loss = self.__attack_loss(preds, data.y)
+            # print(f"iteration: {t}, loss: {loss:.4f}")
+            progress_bar.set_postfix({"Loss": f"{loss:.4f}"})
+
+            # backpropagation of gradients
+            optimizer.zero_grad()
+            loss.backward()
+
+            # Update M
+            optimizer.step()
+
+            with torch.no_grad():
+                M.copy_(projection(M))
+        # ---------------- End ----------------
+
+        # Random Sampling
+        random_sampling = RandomSampling(K=self.num_rand_trials,
+                                         eps=eps,
+                                         A=A,
+                                         attack_loss=self.__attack_loss,
+                                         model=model,
+                                         data=data)
+        M_binary = random_sampling(M)
+        A_pert_binary = A - A * M_binary
+
+        # Convert adjacency matrix to edge list
+        edge_index, _ = dense_to_sparse(A_pert_binary)
+
+        gen_dataset.data.edge_index = edge_index
+        return gen_dataset
+
+    @staticmethod
+    # TODO функция attack_loss должна совпадать с фунцкией потерь, используемой в процессе обучения модели
+    #  (предложение авторов статьи). Поэтому следует расширить функционал атаки для различных loss функций
+    def __attack_loss(preds, labels):
+        return -F.cross_entropy(preds, labels)
 
 
 class NettackEvasionAttacker(EvasionAttacker):
